Extend and clean matching/base.py
