+++ tests/ma/matching/test_match_monthly.py
@@ -6,10 +6,8 @@ import plotly.graph_objects as go
 import data.register
 from ma.matching.match_monthly import (
     MatchMonthly,
-    calculate_matching_score,
-    calculate_supply_surplus_deficit,
-    make_match_monthly,
 )
+from ma.matching.utils import calculate_matching_score, calculate_supply_surplus_deficit
 from ma.ofgem.regos import RegosByTechMonthHolder
 from ma.retailer.consumption import ConsumptionMonthly
 
@@ -28,7 +26,7 @@ def test_calculate_matching_score() -> None:
 def setup() -> Tuple[RegosByTechMonthHolder, ConsumptionMonthly, MatchMonthly]:
     supply = RegosByTechMonthHolder(data.register.REGOS_BY_TECH_MONTH_HOLDER)
     consumption = ConsumptionMonthly(data.register.CONSUMPTION_BY_MONTH)
-    return supply, consumption, make_match_monthly(consumption=consumption, supply=supply)
+    return supply, consumption, MatchMonthly.make(consumption=consumption, supply=supply)
 
 
 def test_match_monthly() -> None:
